pyproject.toml
[tool.poetry]
name = "django-rest-tasks"
version = "0.1.0"
description = ""
authors = ["DevMain1953 <alexanderbrenenburgh@gmail.com>"]
readme = "README.md"
packages = [{include = "django_rest_tasks"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.4"
djangorestframework = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
